examples/flight-sql/python/requirements.txt
adbc_driver_manager
adbc_driver_flightsql
flightsql-dbapi
kamu
sqlalchemy
pandas
pyarrow
jpype1
